Error handeling by filter

+++ src/post/post.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, HttpCode, HttpStatus, Injectable, InternalServerErrorException, Scope } from '@nestjs/common';
+import { BadRequestException, HttpCode, HttpStatus, Injectable, InternalServerErrorException, NotFoundException, Scope } from '@nestjs/common';
 import { UserProvider } from 'src/user/user.provider';
 import { EntityManager, Transaction } from 'typeorm';
 import { Post } from './post.entity';
@@ -14,51 +14,45 @@ export class PostService {
 
     async postFile(file: any, postData: any) {
         let resp: any
-        try {
-            const userId = this.userProvider.getCurrentUser().userId;
-            const filePath: string = file.path;
-            let post = new Post();
-            let postDescription = postData.postDescription.replace(/^'(.*)'$/, '$1');
 
-            await this.entityManager.transaction(async transactionalEntityManager => {
-                post.userId = userId;
-                post.postDescription = postDescription;
-                post.media = filePath;
+        const userId = this.userProvider.getCurrentUser().userId;
+        const filePath: string = file.path;
+        let post = new Post();
+        let postDescription = postData.postDescription.replace(/^'(.*)'$/, '$1');
 
-                await transactionalEntityManager.save(Post, post);
-            });
+        await this.entityManager.transaction(async transactionalEntityManager => {
+            post.userId = userId;
+            post.postDescription = postDescription;
+            post.media = filePath;
 
-            resp = post;
+            await transactionalEntityManager.save(Post, post);
+        });
 
-        } catch (error) {
-            throw error;
-        };
+        resp = post;
+
         return resp;
     };
 
     async getUserPosts() {
         let resp: any;
-        try {
-            const userId = this.userProvider.getCurrentUser().userId;
 
-            const posts = await this.entityManager
-                .createQueryBuilder(Post, 'post')
-                .where('post.userId = :userId', { userId })
-                .andWhere('post.archived = :archived', { archived: false })
-                .getMany();
+        const userId = this.userProvider.getCurrentUser().userId;
+
+        const posts = await this.entityManager
+            .createQueryBuilder(Post, 'post')
+            .where('post.userId = :userId', { userId })
+            .andWhere('post.archived = :archived', { archived: false })
+            .getMany();
 
 
-            for (const post of posts) {
-                const pathParts = post.media.split(/[\/\\]/);
-                // this needs to be changed later so we use only one controller
-                post.media = `${process.env.DOMAIN_NAME}/post/display/${pathParts[pathParts.length - 3]}/${pathParts[pathParts.length - 1]}`;
-            }
-
-            resp = posts;
-
-        } catch (error) {
-            throw error;
+        for (const post of posts) {
+            const pathParts = post.media.split(/[\/\\]/);
+            // this needs to be changed later so we use only one controller
+            post.media = `${process.env.DOMAIN_NAME}/post/display/${pathParts[pathParts.length - 3]}/${pathParts[pathParts.length - 1]}`;
         }
+
+        resp = posts;
+
         return resp;
     };
 
@@ -71,26 +65,21 @@ export class PostService {
     async archivePost(postId: number): Promise<{ message: string; status: number }> {
         let resp: { message: string; status: number };
 
-        try {
-            const userId: string = this.userProvider.getCurrentUser()?.userId;
+        const userId: string = this.userProvider.getCurrentUser()?.userId;
 
-            const result = await this.entityManager
-                .createQueryBuilder()
-                .update(Post)
-                .set({ archived: true })
-                .where('postId = :postId', { postId })
-                .andWhere('userId = :userId', { userId })
-                .andWhere('archived = :archived', { archived: false })
-                .execute();
+        const result = await this.entityManager
+            .createQueryBuilder()
+            .update(Post)
+            .set({ archived: true })
+            .where('postId = :postId', { postId })
+            .andWhere('userId = :userId', { userId })
+            .andWhere('archived = :archived', { archived: false })
+            .execute();
 
-            if (result.affected === 1) {
-                resp = { message: 'postSuccessfullyArchived', status: HttpStatus.OK };
-            } else {
-                resp = { message: 'postIsAlreadyArchivedOrNotExist', status: HttpStatus.OK };
-            }
-
-        } catch (error) {
-            throw error;
+        if (result.affected === 1) {
+            resp = { message: 'postSuccessfullyArchived', status: HttpStatus.OK };
+        } else {
+            resp = { message: 'postIsAlreadyArchivedOrNotExist', status: HttpStatus.OK };
         }
         return resp;
     }
@@ -99,41 +88,40 @@ export class PostService {
     async deletePost(postId: number): Promise<{ message: string; status: number }> {
         let resp: { message: string; status: number };
 
-        try {
-            const userId: string = this.userProvider.getCurrentUser()?.userId;
 
-            const post = await this.entityManager
-                .createQueryBuilder(Post, 'post')
-                .where('post.userId = :userId', { userId })
-                .where('post.postId = :postId', { postId })
-                .getOne();
+        const userId: string = this.userProvider.getCurrentUser()?.userId;
 
+        const post = await this.entityManager
+            .createQueryBuilder(Post, 'post')
+            .where('post.userId = :userId', { userId })
+            .where('post.postId = :postId', { postId })
+            .getOne();
 
-            if (post?.media) {
-                const filePath = path.resolve(post.media);
-                fs.unlink(filePath, (err) => {
-                    if (err) {
-                        throw new InternalServerErrorException("issueDeletingPost");
-                    }
-                });
-            }
+        if (!post) {
+            throw new NotFoundException('postNotFound');
+        }
 
-            const deleteQuery = await this.entityManager
-                .createQueryBuilder()
-                .delete()
-                .from(Post)
-                .where('postId = :postId', { postId })
-                .andWhere('userId = :userId', { userId })
-                .execute();
+        if (post?.media) {
+            const filePath = path.resolve(post.media);
+            fs.unlink(filePath, (err) => {
+                if (err) {
+                    throw new InternalServerErrorException("issueDeletingPost");
+                }
+            });
+        }
 
-            if (deleteQuery.affected === 1 && post.media) {
-                resp = { message: 'postSuccessfullyDeleted', status: HttpStatus.OK };
-            } else {
-                resp = { message: 'postIsAlreadyDeleted', status: HttpStatus.OK };
-            }
+        const deleteQuery = await this.entityManager
+            .createQueryBuilder()
+            .delete()
+            .from(Post)
+            .where('postId = :postId', { postId })
+            .andWhere('userId = :userId', { userId })
+            .execute();
 
-        } catch (error) {
-            throw error;
+        if (deleteQuery.affected === 1 && post.media) {
+            resp = { message: 'postSuccessfullyDeleted', status: HttpStatus.OK };
+        } else {
+            resp = { message: 'postIsAlreadyDeleted', status: HttpStatus.OK };
         }
         return resp;
     }
@@ -141,41 +129,40 @@ export class PostService {
     async editPost(postId: number, postData: EditPostDto): Promise<{ message: string; status: number }> {
 
         let resp: { message: string; status: number };
-        try {
-            const userId: string = this.userProvider.getCurrentUser()?.userId;
 
-            const post = await this.entityManager
-                .createQueryBuilder(Post, 'post')
-                .where('post.userId = :userId', { userId })
-                .where('post.postId = :postId', { postId })
-                .getOne();
+        const userId: string = this.userProvider.getCurrentUser()?.userId;
 
-            const { currPostDesc } = postData;
+        const post = await this.entityManager
+            .createQueryBuilder(Post, 'post')
+            .where('post.userId = :userId', { userId })
+            .where('post.postId = :postId', { postId })
+            .getOne();
 
-            if (currPostDesc === post?.postDescription) {
-                throw new BadRequestException('newDescriptionShouldBeAdded')
-            }
-
-            // for the moment in Post only one description can be edited
-
-            const editPost = await this.entityManager
-                .createQueryBuilder()
-                .update(Post)
-                .set({ postDescription: currPostDesc })
-                .where('postId = :postId', { postId })
-                .andWhere('userId = :userId', { userId })
-                .execute();
-
-
-            if (editPost?.affected === 1) {
-                resp = { message: 'postSuccessfullyEdited', status: HttpStatus.OK };
-            } else {
-                throw new InternalServerErrorException('issueFacedUpdatingPost')
-            }
-
-        } catch (error) {
-            throw error;
+        if (!post) {
+            throw new NotFoundException('postNotFound');
         }
+
+        const { currPostDesc } = postData;
+
+        if (currPostDesc === post?.postDescription) {
+            throw new NotFoundException('newDescriptionShouldBeAdded')
+        }
+
+        const editPost = await this.entityManager
+            .createQueryBuilder()
+            .update(Post)
+            .set({ postDescription: currPostDesc })
+            .where('postId = :postId', { postId })
+            .andWhere('userId = :userId', { userId })
+            .execute();
+
+
+        if (editPost?.affected === 1) {
+            resp = { message: 'postSuccessfullyEdited', status: HttpStatus.OK };
+        } else {
+            throw new InternalServerErrorException('issueFacedUpdatingPost')
+        }
+
         return resp;
     }
 }
